refactor: made all exporter implementation private
