+++ cognite/neat/rules/exporter/_core/rules2labels.py
