+++ cognite/neat/rules/exporter/_rules2triples.py
