+++ cognite/neat/rules/exporter/rules2pydantic_models.py
@@ -17,8 +17,8 @@ from cognite.neat.graph.loaders.core.rdf_to_assets import NeatMetadataKeys
 from cognite.neat.graph.transformations.query_generator.sparql import build_construct_query, triples2dictionary
 from cognite.neat.rules import exceptions
 from cognite.neat.rules.analysis import define_class_asset_mapping, to_class_property_pairs
+from cognite.neat.rules.exporter._rules2dms import DataModel
 from cognite.neat.rules.exporter._validation import are_entity_names_dms_compliant
-from cognite.neat.rules.exporter.rules2dms import DataModel
 from cognite.neat.rules.models.rules import Property, Rules
 from cognite.neat.rules.type_mapping import type_to_target_convention
 from cognite.neat.utils.utils import generate_exception_report
