+++ cognite/neat/workflows/steps/data_contracts.py
@@ -5,7 +5,7 @@ from cognite.client.data_classes import Asset, AssetUpdate, Relationship, Relati
 from cognite.client.data_classes.data_modeling import EdgeApply, NodeApply
 
 from cognite.neat.graph.stores import NeatGraphStore
-from cognite.neat.rules.exporter.rules2dms import DataModel
+from cognite.neat.rules.exporter._rules2dms import DataModel
 from cognite.neat.rules.models.rules import Rules
 from cognite.neat.workflows.steps.step_model import DataContract
