+++ cognite/neat/workflows/steps/lib/graph_extractor.py
@@ -11,7 +11,7 @@ from rdflib import RDF, XSD, Literal, Namespace, URIRef
 from cognite.neat.constants import PREFIXES
 from cognite.neat.graph import extractors
 from cognite.neat.graph.extractors.mocks.graph import generate_triples as generate_mock_triples
-from cognite.neat.rules.exporter.rules2triples import get_instances_as_triples
+from cognite.neat.rules.exporter._rules2triples import get_instances_as_triples
 from cognite.neat.utils.utils import create_sha256_hash
 from cognite.neat.workflows._exceptions import StepNotInitialized
 from cognite.neat.workflows.model import FlowMessage, StepExecutionStatus
